+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Merged from 1.0
 
+* Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 
 ## Version 3.0.0
@@ -24,6 +25,7 @@
 
 ### Merged from 1.0
 
+* Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 * Fix malformed IPv6 for JMX - Issue #306
 
@@ -146,6 +148,7 @@
 
 ### Merged from 1.0
 
+* Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 * Fix malformed IPv6 for JMX - Issue #306
 * Step karaf to 4.2.8
@@ -156,6 +159,7 @@
 
 #### Merged from 1.0
 
+* Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 * Fix malformed IPv6 for JMX - Issue #306
 
@@ -193,6 +197,7 @@
 
 ## Version 1.0.8 (Not yet released)
 
+* Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 * Fix malformed IPv6 for JMX - Issue #306
 * Step karaf to 4.2.8