+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## Version 4.0.5 (Not yet released)
 
+### Merged from 1.0
+
+* Skip unnecessary reads from repair history - Issue #548
+
 ## Version 4.0.4
 
 ### Merged from 1.0
@@ -54,6 +58,7 @@
 
 ### Merged from 1.0
 
+* Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 
 ## Version 3.0.0
@@ -74,6 +79,7 @@
 
 ### Merged from 1.0
 
+* Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 * Fix malformed IPv6 for JMX - Issue #306
 
@@ -196,6 +202,7 @@
 
 ### Merged from 1.0
 
+* Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 * Fix malformed IPv6 for JMX - Issue #306
 * Step karaf to 4.2.8
@@ -206,6 +213,7 @@
 
 #### Merged from 1.0
 
+* Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 * Fix malformed IPv6 for JMX - Issue #306
 
@@ -243,6 +251,7 @@
 
 ## Version 1.0.8 (Not yet released)
 
+* Skip unnecessary reads from repair history - Issue #548
 * Fix repair job priority - Issue #515
 * Fix malformed IPv6 for JMX - Issue #306
 * Step karaf to 4.2.8